backend/requirements.txt
fastapi
uvicorn
llama-index
llama-index-embeddings-openai
llama-index-llms-openai
llama-index-vector-stores-postgres
llama-index-postprocessor-sentence-transformer-rerank
sentence-transformers
sqlalchemy
asyncpg
python-dotenv
psycopg2-binary
pgvector
python-multipart
databases[postgresql]
